requirements.txt
Flask==3.0.0
flask-cors==4.0.0
supabase==2.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.32.3
psycopg[binary]==3.1.18

